backend/requirements.txt
fastapi[standard]>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.0
langchain>=0.2.0
langchain-google-genai>=1.0.4
langgraph>=0.0.60
supabase>=2.4.5
python-dotenv>=1.0.1
vecs>=0.4.0
python-multipart>=0.0.9
